requirements.txt
discord.py>=2.7.0
python-dotenv==1.0.1
groq>=1.0.0
yt-dlp>=2025.2.1
PyNaCl==1.5.0
davey
youtube-search
aiosqlite==0.20.0
aiohttp==3.9.5
APScheduler==3.10.4
Pillow==10.3.0
wavelink==3.3.0
spotipy==2.23.0
lyricsgenius==3.0.1
psutil==5.9.8
colorlog==6.8.2
